backend/requirements.txt
# FastAPI core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Pydantic v2
pydantic>=2.6.0
pydantic-settings>=2.2.0

# MongoDB async
motor>=3.4.0
pymongo>=4.6.0

# OCR
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0

# LangChain + OpenAI (optional — only needed if LLM_ENABLED=true)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.30.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.2.0
python-dateutil>=2.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
aiofiles>=23.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
